data-plane/model-server/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Cloud storage
boto3==1.28.85
google-cloud-storage==2.14.0

# ML frameworks
numpy==1.24.3
scikit-learn==1.3.2

# PyTorch (optional - uncomment if needed)
# torch==2.1.0
